Create PostService interface for API calls
This commit introduces the `PostService` interface, which defines the API endpoints for interacting with posts.

It includes a suspend function `getPosts()` annotated with `@GET("/posts")` to fetch a list of posts from the "/posts" endpoint. The function returns a `Response<Posts>`.

Additionally, the `Post` data class was modified by removing an unnecessary newline at the end of the file.